+++ huaweicloud/doc.go
@@ -1,0 +1,2 @@
+// Package huaweicloud provides the implementation of the Import from Huawei Cloud.
+package huaweicloud